Create LICENSE
MIT License baby, I hope you all enjoy and use it ethically.  I have been blessed by Jesus to be able to have a version on here that I'm of free-will to distribute to the public in hopes that this brings a more secure internet, and promotes Freedom of thought into the future as well as promotes Democracy for all.